Paging implementation extended Can create new page directories now and map physical address ranges to virtual addresses.

+++ second_stage/paging.h
@@ -8,10 +8,10 @@
 #include <stdint.h>
 
 //! Number of entries within a page table/directory.
-#define PAGE_ENTRY_COUNT (1 << 10)
+#define PAGE_ENTRY_COUNT (uint32_t)(1 << 10)
 
 //! Size of a single page or page table/directory.
-#define PAGE_SIZE (1 << 12)
+#define PAGE_SIZE (uint32_t)(1 << 12)
 
 //! Represents a single page entry in a page table.
 typedef union {
